feat(web): はじめにセクションに Overview を追加し /docs のデフォルト表示にする
「はじめに」セクションの先頭に「概要」（Overview）ナビ項目を追加し、その内容
（getting-started/index.mdx）を /docs のデフォルトコンテンツとして表示する。
URL を二重化させないため、この項目の href は /docs/getting-started ではなく
/docs 自体に統一する。

- 共有コンポーネント DocArticle を features/docs に切り出し、本文・パンくず・
  関連記事・ページ送り・右側目次（blog と共通の2card）を [...slug] と /docs の
  両方で再利用する
- navigation.ts: getting-started の index 項目だけ title を「概要」、href を
  /docs にする
- /docs: getMDXContentForRSC('getting-started', locale) が見つかれば Overview を
  表示。コンテンツが無いロケール（例: 現状の en）は従来のセクション一覧に
  フォールクバックする
