+++ apps/web/messages/en/common.json
@@ -283,6 +283,7 @@
     "landing": {
       "subtitle": "Get started with Dayopt using our documentation and guides"
     },
+    "overview": "Overview",
     "categories": {
       "getting-started": "Getting Started",
       "features": "Features",
