+++ apps/web/messages/ja/common.json
@@ -252,6 +252,7 @@
     "landing": {
       "subtitle": "ドキュメントとガイドでDayoptを始めましょう"
     },
+    "overview": "概要",
     "categories": {
       "getting-started": "はじめに",
       "features": "機能",
