+++ apps/web/src/app/[locale]/(docs)/docs/[...slug]/page.tsx
@@ -1,42 +1,11 @@
-import { FAQStructuredData } from '@/components/seo/EnhancedSEO';
-import { Breadcrumbs, mdxComponents, PageNavigation, TableOfContentsCards } from '@/features/docs';
-import { faqMdxComponents } from '@/features/docs/components/MDXComponents';
-import { getAllContent, getMDXContentForRSC, getRelatedContent } from '@/lib/mdx';
+import { DocArticle } from '@/features/docs';
+import { getAllContent, getMDXContentForRSC } from '@/lib/mdx';
 import { Link } from '@/platform/i18n/navigation';
 import { ContentData } from '@/types/content';
-import { Card, CardHeader, CardTitle, Heading, Text } from '@dayopt/components';
+import { Heading, Text } from '@dayopt/components';
 import { Metadata } from 'next';
 import { getTranslations } from 'next-intl/server';
-import { MDXRemote } from 'next-mdx-remote/rsc';
 import { notFound } from 'next/navigation';
-import rehypeHighlight from 'rehype-highlight';
-import remarkGfm from 'remark-gfm';
-
-/**
- * MDXソースからFAQ（h2 = 質問、本文 = 回答）を抽出
- */
-function extractFAQsFromMDX(content: string): Array<{ question: string; answer: string }> {
-  const faqs: Array<{ question: string; answer: string }> = [];
-  const lines = content.split('\n');
-  let currentQuestion = '';
-  let currentAnswer = '';
-
-  for (const line of lines) {
-    if (line.startsWith('## ')) {
-      if (currentQuestion && currentAnswer) {
-        faqs.push({ question: currentQuestion, answer: currentAnswer.trim() });
-      }
-      currentQuestion = line.replace('## ', '');
-      currentAnswer = '';
-    } else if (currentQuestion && line.trim()) {
-      currentAnswer += line + ' ';
-    }
-  }
-  if (currentQuestion && currentAnswer) {
-    faqs.push({ question: currentQuestion, answer: currentAnswer.trim() });
-  }
-  return faqs;
-}
 
 interface PageParams {
   locale: string;
@@ -148,7 +117,7 @@ export default async function DocPage({ params }: DocPageProps) {
 
   try {
     const slug = slugArray.join('/');
-    const category = slugArray[0];
+    const category = slugArray[0] ?? '';
     const contentSlug = slugArray.slice(1).join('/');
 
     // Get MDX content
@@ -173,85 +142,19 @@ export default async function DocPage({ params }: DocPageProps) {
     }
 
     const { content: mdxContent, frontMatter } = content;
-
-    // Get adjacent pages and related content in parallel
-    const [{ previousPage, nextPage }, relatedContent] = await Promise.all([
-      getAdjacentPages(slug, locale),
-      getRelatedContent(frontMatter.category, slug, 3, locale),
-    ]);
+    const { previousPage, nextPage } = await getAdjacentPages(slug, locale);
 
     return (
-      <div className="flex">
-        {/* Main Content */}
-        <div className="min-w-0 flex-1 px-6 py-8 lg:px-8">
-          <div className="mx-auto max-w-3xl">
-            {/* Breadcrumb navigation */}
-            <Breadcrumbs slug={slug} title={frontMatter.title} />
-
-            {/* FAQ構造化データ（FAQサブページのみ） */}
-            {category === 'faq' &&
-              contentSlug &&
-              (() => {
-                const faqs = extractFAQsFromMDX(mdxContent);
-                return faqs.length > 0 ? <FAQStructuredData faqs={faqs} /> : null;
-              })()}
-
-            {/* MDX content */}
-            <article>
-              <MDXRemote
-                source={mdxContent}
-                components={category === 'faq' && contentSlug ? faqMdxComponents : mdxComponents}
-                options={{
-                  mdxOptions: {
-                    remarkPlugins: [remarkGfm],
-                    rehypePlugins: [rehypeHighlight],
-                  },
-                }}
-              />
-            </article>
-
-            {/* Tags section */}
-            {/* Related content */}
-            {relatedContent.length > 0 && (
-              <aside className="border-border mt-12 border-t pt-8">
-                <Heading as="h2" size="xl" className="mb-6">
-                  {tDocs('relatedArticles')}
-                </Heading>
-                <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3">
-                  {relatedContent.map((related) => (
-                    <Link key={related.slug} href={`/docs/${related.slug}`} className="block">
-                      <Card className="hover:bg-state-hover h-full gap-4 py-4 transition-colors">
-                        <CardHeader className="gap-2 px-4 py-0">
-                          <CardTitle className="line-clamp-2 text-sm">
-                            {related.frontMatter.title}
-                          </CardTitle>
-                          <div className="flex items-center gap-2">
-                            {related.frontMatter.updatedAt && (
-                              <time className="text-muted-foreground text-xs">
-                                {new Date(related.frontMatter.updatedAt).toLocaleDateString(locale)}
-                              </time>
-                            )}
-                          </div>
-                        </CardHeader>
-                      </Card>
-                    </Link>
-                  ))}
-                </div>
-              </aside>
-            )}
-
-            {/* Previous/next page navigation */}
-            <PageNavigation previousPage={previousPage} nextPage={nextPage} />
-          </div>
-        </div>
-
-        {/* Right Sidebar - Table of Contents（xl以上で表示。blog と共通の2card レイアウト） */}
-        <aside className="hidden w-72 flex-shrink-0 xl:block">
-          <div className="sticky top-14 py-8 pr-6 lg:pr-8">
-            <TableOfContentsCards content={mdxContent} />
-          </div>
-        </aside>
-      </div>
+      <DocArticle
+        slug={slug}
+        category={category}
+        contentSlug={contentSlug}
+        mdxContent={mdxContent}
+        frontMatter={frontMatter}
+        locale={locale}
+        previousPage={previousPage}
+        nextPage={nextPage}
+      />
     );
   } catch {
     // Error page
