+++ apps/web/src/features/docs/index.ts
@@ -3,6 +3,7 @@ export { Breadcrumbs } from './components/Breadcrumbs';
 export { ClientSidebar } from './components/ClientSidebar';
 export { ClientTableOfContents } from './components/ClientTableOfContents';
 export { CopyCodeButton } from './components/CopyCodeButton';
+export { DocArticle } from './components/DocArticle';
 export { faqMdxComponents, mdxComponents } from './components/MDXComponents';
 export { PageNavigation } from './components/PageNavigation';
 export { TableOfContentsCards } from './components/TableOfContentsCards';
